Fixing minor bugs & Refactoring
