+++ Meeting-App/app.js
@@ -1,12 +1,12 @@
 require('dotenv').config({ path: __dirname + '/.env' });
-require('./src/config/dbConn');
-require('./src/config/sequelizeConn');
+require('./src/config/db-config');
+require('./src/config/sequelize-config');
 
 const express = require('express');
 const app = express();
 
-const userRoutes = require('./src/routes/usersRoute');
-const meetingRoutes = require('./src/routes/meetingsRoute');
+const userRoutes = require('./src/routes/users-route');
+const meetingRoutes = require('./src/routes/meetings-route');
 
 const PORT = process.env.PORT;
