+++ Meeting-App/src/models/meetings-model.js
@@ -1,5 +1,5 @@
 const { DataTypes } = require('sequelize');
-const sequelize = require('../config/sequelizeConn');
+const sequelize = require('../config/sequelize-config');
 
 const Meeting = sequelize.define('Meeting', {
     title: {
