+++ Meeting-App/src/models/usersModel.js
@@ -1,5 +1,5 @@
 const { DataTypes } = require('sequelize');
-const sequelize = require('../config/sequelizeConn');
+const sequelize = require('../config/sequelize-config');
 
 const User = sequelize.define('User', {
     name: {
